Extracted lifecycle to own ScannerStateManager
